+++ src/lib.rs
@@ -14,3 +14,5 @@ mod tests;
 mod native;
 mod proto_source;
 mod routes;
+
+pub mod region;